only install clang-format if it is really missing
As the default image already has v10, v11 (default) and v12 installed,
you don't have to install v8 and don't even use it. This removes the
need for a long (most of the time 2 minutes) mirror search or even call
'apt-get install' in the first place. This shortens the time spent in
actions_install.sh from about 2:30min to 6s, which is 25x faster.